Much better structure, working on login flow
