+++ lib/screens/login_screen.dart
@@ -3,6 +3,9 @@ import 'package:aad_oauth/model/config.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
+import '../widgets/disco_bar.dart';
+import '../widgets/ht_banner.dart';
+
 final navigatorKey = GlobalKey<NavigatorState>();
 
 class LoginScreen extends StatefulWidget {
@@ -20,84 +23,39 @@ class _LoginScreenState extends State<LoginScreen> {
       navigatorKey: navigatorKey,
       loader: const SizedBox());
 
-  late Future<bool> loggedin;
-  final Future<SharedPreferences> _prefs = SharedPreferences.getInstance();
-
-  @override
-  void initState() {
-    super.initState();
-    loggedin = _prefs.then((SharedPreferences prefs) {
-      prefs.remove("loggedin");
-      return prefs.getBool('loggedin') ?? false;
-    });
-  }
-
   @override
   Widget build(BuildContext context) {
-    return FutureBuilder<bool>(
-        future: loggedin,
-        builder: (BuildContext context, AsyncSnapshot<bool> snapshot) {
-          switch (snapshot.connectionState) {
-            case ConnectionState.none:
-            case ConnectionState.waiting:
-              return const CircularProgressIndicator();
-            case ConnectionState.active:
-            case ConnectionState.done:
-              if (snapshot.hasError) {
-                return Text('Error: ${snapshot.error}');
-              } else {
-                if (snapshot.data == true) {
-                  Future(() {
-                    if (context.mounted) {
-                      Navigator.pushNamedAndRemoveUntil(
-                          context, "/home", (_) => false);
-                    }
-                  });
-                  return const CircularProgressIndicator();
-                } else {
-                  return Scaffold(
-                    appBar: AppBar(
-                      backgroundColor: Theme.of(context).colorScheme.primary,
-                      title: Text(
-                        "login",
-                        style: Theme.of(context).textTheme.displayMedium,
-                      ),
-                    ),
-                    body: ListView(
-                      children: <Widget>[
-                        SizedBox(
-                          height: 100,
-                          width: 300,
-                          child: Card(
-                            shape: RoundedRectangleBorder(
-                                borderRadius: BorderRadius.circular(0.0)),
-                            elevation: 5,
-                            color: Theme.of(context).colorScheme.primary,
-                            child: ListTile(
-                              minVerticalPadding: 10.0,
-                              title: Text(
-                                "hendrix today",
-                                style:
-                                    Theme.of(context).textTheme.displayMedium,
-                                textAlign: TextAlign.center,
-                              ),
-                            ),
-                          ),
-                        ),
-                        ListTile(
-                          leading: Icon(Icons.launch),
-                          title: Text('Login'),
-                          onTap: () {
-                            login();
-                          },
-                        ),
-                      ],
-                    ),
-                  );
-                }
-              }
-          }
-        });
+    return Scaffold(
+      body: Column(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+        children: [
+          const HTBanner(),
+          SizedBox(
+            height: 62,
+            width: 200,
+            child: Card(
+              shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.circular(0.0)),
+              elevation: 5,
+              color: Theme.of(context).colorScheme.primary,
+              child: ListTile(
+                minVerticalPadding: 10.0,
+                title: Text(
+                  "login",
+                  style: Theme.of(context).textTheme.displayMedium,
+                  textAlign: TextAlign.center,
+                ),
+                leading: Icon(Icons.launch,
+                    size: 30, color: Theme.of(context).iconTheme.color),
+                onTap: () => login(),
+              ),
+            ),
+          ),
+          const DiscoBar(),
+        ],
+      ),
+    );
   }
 
   void showError(dynamic ex) {
